Various small code cleanups: - Lots of "signed vs unsigned" comparison fixes - time_t globally replaced with TIME_T - All seqnos moved to "seqno_t", which defaults to unsigned long - DIMOF() definition centralized to portability.h and typecast to int - EOS define moved to portability.h - dropped inclusion of signal.h from stonith.h, so that sigignore is   properly defined
CVS patchset: 1584
CVS date: 2003/01/31 10:02:09

+++ heartbeat/client_lib.c
@@ -1629,7 +1629,7 @@ sendnodemsg(ll_cluster_t* lcl, struct ha_msg* msg
 }
 
 static char	APILogBuf[MAXLINE] = "";
-int		BufLen = 0;
+size_t		BufLen = 0;
 
 static void
 ClearLog(void)
@@ -1648,7 +1648,7 @@ APIError(ll_cluster_t* lcl)
 static void
 ha_api_log(int priority, const char * fmt, ...)
 {
-	int	len;
+	size_t	len;
         va_list ap;
         char buf[MAXLINE];
